moved to stereosgbm module
