+++ stereosgbm/__init__.py
@@ -1,1 +1,3 @@
 from .disparity_calculator import DisparityCalculator
+from .disparity_calculator import EdgeBasedDisparityCalculator
+
